% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkstudio.R
\name{activate}
\alias{activate}
\title{Activate daRkStudio}
\usage{
activate(path = NULL, backup = TRUE)
}
\arguments{
\item{path}{character:
Path to RStudio's \code{index.htm}. Useful for times when the default
installation method cannot successfully locate the file.}

\item{backup}{logical:
  TRUE or FALSE. Copies the default \code{index.htm} file to
  \code{index.htm.pre-ds}. Defaults to TRUE.

daRkStudio modifies \code{index.htm}, a file used by RStudio to construct
it's DOM (Document Object Model).

The only change to \code{index.htm} the inclusion of a \code{<link>} element
near the end of the file, which tells RStudio to load \code{darkstudio.css}.

daRkStudio creates a directory, "darkstudio", inside the \code{www} folder,
which is found at \code{/Applications/RStudio.app/Contents/Resources/www} on
macOS, and \code{C:\\Program Files\\RStudio\\www} on Windows.

\code{activate()} will create a backup of \code{index.htm} at
\code{www/darkstudio/index.htm.pre-ds}. \code{www/darkstudio} is also where
you will find \code{darkstudio.css}, which is the bread and butter of this
package.

On Windows, you will likely need Administrator Privileges if you've
installed RStudio to the default location, \code{C:\\Program Files\\RStudio}.}
}
\value{
TRUE

Returns \code{TRUE} if the operation is successful.
}
\description{
Activate daRkStudio
}
\examples{
\dontrun{
# Default:
activate()

# macOS:
path_index <- "/Applications/RStudio.app/Contents/Resources/www/index.htm"
activate(path = path_index, backup = TRUE)

# Windows:
path_index <- "C:/Program Files/RStudio/www/index.htm"
activate(path = path_index, backup = TRUE)
}

}
