% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customise.R
\name{customise}
\alias{customise}
\title{custom.R
Created by Stefano Coretta on 2023-12-18}
\usage{
customise(ds_css, font_family)
}
\description{
custom.R
Created by Stefano Coretta on 2023-12-18
}
\keyword{internal}
