% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkstudio-package.R
\docType{package}
\name{darkstudio-package}
\alias{darkstudio}
\alias{darkstudio-package}
\title{darkstudio: An Alternative Look for RStudio's Modern and Sky Themes}
\description{
Modifies RStudio's Modern and Sky themes to use grey rather than blue.
}
\details{
Documentation for darkstudio
}
\keyword{internal}
