% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deactivate.R
\name{deactivate}
\alias{deactivate}
\title{Deactivate daRkStudio}
\usage{
deactivate(path = NULL, restore_index = TRUE)
}
\arguments{
\item{path}{character:
Path to RStudio's \code{index.htm}.}

\item{restore_index}{boolean:
restore initial \code{index.htm} file from a created backup.}
}
\value{
Returns \code{TRUE} if the operation is successful.
}
\description{
Remove and replace the modified \code{index.htm} with the backup
\code{index.htm.pre-ds} file. Also deletes the \code{darkstudio} directory
as well as the \code{inverse.js} script.
}
\details{
This function does NOT uninstall the daRkStudio package.
To uninstall the darkstudio package, copy and
paste remove.packages('darkstudio') into the console.
}
