% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\docType{data}
\name{index}
\alias{index}
\title{index.R
Created by Riley Roach on 2020-05-24}
\format{
An object of class \code{list} of length 6.
}
\usage{
index
}
\description{
index.R
Created by Riley Roach on 2020-05-24
}
\keyword{internal}
