% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_file_find}
\alias{index_file_find}
\title{index.R
Created by Riley Roach on 2020-05-24}
\usage{
index_file_find(path = NULL)
}
\description{
index.R
Created by Riley Roach on 2020-05-24
}
\keyword{internal}
