% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{preview_changes}
\alias{preview_changes}
\title{tools.R
 Created by Riley Roach on 2020-05-30}
\usage{
preview_changes(path = ".")
}
\description{
tools.R
 Created by Riley Roach on 2020-05-30
}
\keyword{internal}
