% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings_dir_create}
\alias{settings_dir_create}
\title{storage.R
Created by Riley Roach on 2020-05-27}
\usage{
settings_dir_create(path = NULL)
}
\description{
storage.R
Created by Riley Roach on 2020-05-27
}
\keyword{internal}
