% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkstudio.R
\name{uninstall_darkstudio}
\alias{uninstall_darkstudio}
\title{Uninstall daRkStudio}
\usage{
uninstall_darkstudio(index_file = NULL)
}
\arguments{
\item{index_file}{character:
Path to RStudio's \code{index.htm}.}
}
\value{
TRUE
}
\description{
Remove and replace the modified \code{index.htm} with the backup
\code{index.htm.pre-ds} file. Also deletes the \code{darkstudio} directory.
}
\details{
This function does NOT uninstall the daRkStudio package.
To uninstall the darkstudio package, copy and
paste remove.packages('darkstudio') into the console.
}
