% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkstudio.R
\name{update_darkstudio}
\alias{update_darkstudio}
\title{Update darkstudio.css}
\usage{
update_darkstudio()
}
\value{
TRUE
}
\description{
Updates the \code{darkstudio.css}. Meant to be used after upgrading the
daRkStudio package.
}
