% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_styles.R
\name{update_styles}
\alias{update_styles}
\title{Update Styles}
\usage{
update_styles(path = NULL)
}
\value{
Returns \code{TRUE} if the operation is successful.
}
\description{
Updates the \code{darkstudio.css}. Meant to be used after upgrading
\code{daRkStudio} and/or after upgrading RStudio.
}
