% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkstudio.R
\name{uninstall_darkstudio}
\alias{uninstall_darkstudio}
\title{Uninstall daRkStudio}
\usage{
uninstall_darkstudio(index_file = NULL)
}
\value{
nothing
}
\description{
Remove and replace the modified \code{index.htm} with the backup
\code{index.htm.pre-ds} file.
}
